Add test case for renaming a column and only applying inserts
This catches a bug in `sql.BuildDMLInsertQuery` where we we are fetching
the insert values using the renamed column's name, and end up fetching
the value of the wrong column.

The test in `localtests/rename` did not catch this because binlog
update events were "correcting" the error, as they follow a different
code path that does not contain the bug.