Add `onTestReplicaComplete` hook
When doing `--test-on-replica` runs, it's common to want to perform
some sort of integrity checks against the data and schema of your
tables. To cater for this in a more automated means, this adds a new
hook `onTestReplicaComplete` which is fired once the migration is
completed and cutover but before replication is restarted to ensure
consistency with rows.