Builds go inside docker container
This allows us to quickly upgrade (or test) go versions. It also
simplifies the current process that downloads the go binaries and
messing with the GOPATH.

We build in the docker/ folder to avoid sending the entire directory
(vendor/ included) to the docker context. That can add some unneccesary
time to the build since we're not copying anything.