fix returned error values in defer functions
This is copied from https://github.com/github/orchestrator/pull/946 with
a slight tweak to the an indent.

Originates from the PR https://github.com/github/orchestrator/issues/943
that applies to this codebase too.

When we define the variable and attempt to set it in a defer function,
the value we're returning isn't what we expect.

Instead we can name the return values as variables and set that instead.