Adds some support for running gh-ost in `MINIMAL` RBR setups
Partially addresses #230.

Add a new flag, `--switch-to-rbr-full`. If this flag is set, and
`--assume-rbr` is **not** passed, gh-ost will first set binlog format to
`ROW` if it is not already (this is existing behavior), but then _also_
sets the binlog row image to `FULL` if it is not already (this is the
new behavior) before restarting replication.

Mirroring the way gh-ost handles changing binary logging from SBR to
RBR, gh-ost will _not_ reset the row image to `MINIMAL` under any
circumstance.