Revertible Migration (#1607)
This PR introduces revertible migrations, following the approach suggested by @jonahberquist. In case a migration causes production impact after the cut-over, it can be reverted quickly while preserving the writes that happened after the cut-over.

When gh-ost is invoked with the --checkpoint flag and the migration completes, the migration can be reverted by invoking gh-ost again with the --revert flag and the --old-table flag specifying the name of the "old" table from the first migration e.g. _mytable_del. Also see docs/revert.md.